Update README.md
This is NOT a official version of this plugin. This is the fork made for testing purposes. 
Changes have been applied, removing the jQuery and Bootstrap libraries and the Lite version of Summernote editor was used - see to file "summernote-editor.php". Additional icons in the editor menu have been replaced with temporary SVG icons that load directly from the "/js/files.js" file. The editor still needs to adjust the stylesheets in the theme used so that all elements are properly displayed and visible.